Refactored Login Feature to Use React Hook Form and Updated Dependencies
